+++ src/baseq2/g_spawn.c
@@ -26,7 +26,7 @@ typedef struct {
 
 typedef struct {
     char    *name;
-    size_t  ofs;
+    unsigned ofs;
     fieldtype_t type;
 } spawn_field_t;
